Move results client
Puts the results client into the client package.

Closes #331
