+++ pkg/client/results/reader.go
