+++ pkg/client/results/types.go
